+++ webapp-frontend/src/app/admin/user.component.spec.ts
@@ -19,21 +19,21 @@
  */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AdminComponent } from './admin.component';
+import { UserComponent } from './user.component';
 
-describe('AdminComponent', () => {
+describe('UserComponent', () => {
   let component: AdminComponent;
-  let fixture: ComponentFixture<AdminComponent>;
+  let fixture: ComponentFixture<UserComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ AdminComponent ]
+      declarations: [UserComponent]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(AdminComponent);
+    fixture = TestBed.createComponent(UserComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });