+++ docs/release-notes.rst
@@ -39,6 +39,7 @@ Version 1.0.4, 21 June 2019
 * Add build number to dashboard version string
 * Move mock admin screen user data to backend
 * Update App manager client to spec version 0.1.5
+* Rework admin table
 
 Version 1.0.3, 28 May 2019
 --------------------------